+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added async editor that does not depend on std
+
 ## [0.2.0] - 2022-03-22
 
 - Added basic line history